requirements.txt
attrs==22.1.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
config==0.5.1
connexion==2.14.1
Flask==3.0.3
idna==3.6
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.16.0
MarkupSafe==2.1.5
mysql-connector==2.2.9
mysql-connector-python==8.3.0
packaging==21.3
pyparsing==3.0.9
pyrsistent==0.19.1
PyYAML==6.0
requests==2.31.0
swagger-ui-bundle==0.0.9
urllib3==2.2.1
Werkzeug==3.0.2
